LayoutTests/fast/canvas/webgl/copy-tex-image-and-sub-image-2d-bad-input-expected.txt


Verify copyTexImage2D and copyTexSubImage2D can handle bad input

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing with antialias on
Testing copyTexImage2D
PASS getError was expected value: NO_ERROR :
PASS getError was expected value: NO_ERROR :
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Testing copyTexSubImage2D
PASS getError was expected value: NO_ERROR :
PASS getError was expected value: NO_ERROR :
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Testing with antialias off
Testing copyTexImage2D
PASS getError was expected value: NO_ERROR :
PASS getError was expected value: NO_ERROR :
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Testing copyTexSubImage2D
PASS getError was expected value: NO_ERROR :
PASS getError was expected value: NO_ERROR :
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS successfullyParsed is true

TEST COMPLETE
